# Vectorize `basic_string::find_last_of`