--- a/benchmarks/src/find_first_of.cpp
+++ b/benchmarks/src/find_first_of.cpp
@@ -13,7 +13,7 @@
 
 using namespace std;
 
-enum class AlgType : bool { std_func, str_member };
+enum class AlgType { std_func, str_member_first, str_member_last };
 
 template <AlgType Alg, class T, T Start = T{'a'}>
 void bm(benchmark::State& state) {
@@ -22,7 +22,7 @@ void bm(benchmark::State& state) {
     const size_t HSize = Pos * 2;
     const size_t Which = 0;
 
-    using container = conditional_t<Alg == AlgType::str_member, basic_string<T>, vector<T>>;
+    using container = conditional_t<Alg == AlgType::std_func, vector<T>, basic_string<T>>;
 
     container h(HSize, T{'.'});
     container n(NSize, T{0});
@@ -37,8 +37,10 @@ void bm(benchmark::State& state) {
     for (auto _ : state) {
         benchmark::DoNotOptimize(h);
         benchmark::DoNotOptimize(n);
-        if constexpr (Alg == AlgType::str_member) {
-            benchmark::DoNotOptimize(h.find_first_of(n.data(), 0, n.size()));
+        if constexpr (Alg == AlgType::str_member_first) {
+            benchmark::DoNotOptimize(h.find_first_of(n));
+        } else if constexpr (Alg == AlgType::str_member_last) {
+            benchmark::DoNotOptimize(h.find_last_of(n));
         } else {
             benchmark::DoNotOptimize(find_first_of(h.begin(), h.end(), n.begin(), n.end()));
         }
@@ -46,8 +48,8 @@ void bm(benchmark::State& state) {
 }
 
 void common_args(auto bm) {
-    bm->Args({2, 3})->Args({7, 4})->Args({9, 3})->Args({22, 5})->Args({58, 2});
-    bm->Args({102, 4})->Args({325, 1})->Args({1011, 11})->Args({1502, 23})->Args({3056, 7});
+    bm->Args({2, 3})->Args({7, 4})->Args({9, 3})->Args({22, 5})->Args({58, 2})->Args({102, 4});
+    bm->Args({325, 1})->Args({400, 50})->Args({1011, 11})->Args({1502, 23})->Args({3056, 7});
 }
 
 BENCHMARK(bm<AlgType::std_func, uint8_t>)->Apply(common_args);
@@ -55,8 +57,14 @@ BENCHMARK(bm<AlgType::std_func, uint16_t>)->Apply(common_args);
 BENCHMARK(bm<AlgType::std_func, uint32_t>)->Apply(common_args);
 BENCHMARK(bm<AlgType::std_func, uint64_t>)->Apply(common_args);
 
-BENCHMARK(bm<AlgType::str_member, char>)->Apply(common_args);
-BENCHMARK(bm<AlgType::str_member, wchar_t>)->Apply(common_args);
-BENCHMARK(bm<AlgType::str_member, wchar_t, L'\x03B1'>)->Apply(common_args);
+BENCHMARK(bm<AlgType::str_member_first, char>)->Apply(common_args);
+BENCHMARK(bm<AlgType::str_member_first, wchar_t>)->Apply(common_args);
+BENCHMARK(bm<AlgType::str_member_first, wchar_t, L'\x03B1'>)->Apply(common_args);
+BENCHMARK(bm<AlgType::str_member_first, char32_t>)->Apply(common_args);
+BENCHMARK(bm<AlgType::str_member_first, char32_t, L'\x03B1'>)->Apply(common_args);
+
+BENCHMARK(bm<AlgType::str_member_last, char>)->Apply(common_args);
+BENCHMARK(bm<AlgType::str_member_last, wchar_t>)->Apply(common_args);
+BENCHMARK(bm<AlgType::str_member_last, wchar_t, L'\x03B1'>)->Apply(common_args);
 
 BENCHMARK_MAIN();