--- a/stl/inc/__msvc_string_view.hpp
+++ b/stl/inc/__msvc_string_view.hpp
@@ -833,7 +833,7 @@ constexpr size_t _Traits_find_last_of(_In_reads_(_Hay_size) const _Traits_ptr_t<
                 bool _Try_vectorize = false;
 
                 if constexpr (sizeof(_Elem) <= 2) {
-                    _Try_vectorize = _Hay_size - _Start_at > _Threshold_find_first_of;
+                    _Try_vectorize = _Hay_start + 1 > _Threshold_find_first_of;
                     // Additional condition for when the vectorization outperforms the table lookup
                     constexpr size_t _Find_last_of_bitmap_threshold = sizeof(_Elem) == 1 ? 48 : 8;