@StephanTLavavej 2024-09-09
I've pushed a merge with `main` to resolve a trivial adjacent-add conflict in `vector_algorithms.cpp`.