@StephanTLavavej 2024-09-12
Conflict-free merge with toolset update, no clang-format regen needed.